requirements.txt
python-telegram-bot==13.15
APScheduler==3.10.4
beautifulsoup4==4.12.2
requests==2.31.0
